Allow guide to directly modify angular velocity
This is cheap, but very effective for eliminating the bobblehead effect
seen on seeking drones. That is behavior that should move to the drone
steer method. Also need to establish a return object for cases where the
guidance module shouldn't directly call steer and thrust methods. Also,
waht about braking? Steering brake, too?